Some initialization and nudging fixes.
git-svn-id: https://svn.code.sf.net/p/olam-model/code/branches/6.0_merger@875 2bc74c4e-f164-4307-89b2-97abcff1156f

+++ oisan/spec_nudge.f90
@@ -117,11 +117,11 @@ subroutine nudge_prep_spec(iaction, o_rho, o_theta, o_rrw, o_uzonal, o_umerid)
 
   integer, intent(in) :: iaction
 
-  real(r8), intent(in)  :: o_rho   (mza,mwa)
-  real,     intent(in)  :: o_theta (mza,mwa)
-  real,     intent(in)  :: o_rrw   (mza,mwa)
-  real,     intent(in)  :: o_uzonal(mza,mwa)
-  real,     intent(in)  :: o_umerid(mza,mwa)
+  real, intent(in) :: o_rho   (mza,mwa)
+  real, intent(in) :: o_theta (mza,mwa)
+  real, intent(in) :: o_rrw   (mza,mwa)
+  real, intent(in) :: o_uzonal(mza,mwa)
+  real, intent(in) :: o_umerid(mza,mwa)
 
   integer               :: j,iw,k,iwnud,iwnud1
